TST: Reorganize IntervalIndex tests (#28800)
